requirement.txt
ultralytics
opencv-python
pyttsx3
googlemaps
SpeechRecognition
pyaudio
python-dotenv
